Some new tests
